requirements.txt
pydantic==2.7.4
jinja2==3.1.4
requests==2.32.3
python-dotenv==1.0.1
openai>=1.3.0
weasyprint>=62.0
cryptography>=3.4.8